+++ RELEASENOTES.md
@@ -30,6 +30,9 @@
     *   Add new error code
         `PlaybackException.ERROR_CODE_DECODING_RESOURCES_RECLAIMED` that is used
         when codec resources are reclaimed for higher priority tasks.
+    *   Let `AdsMediaSource` load preroll ads before initial content media
+        preparation completes
+        ([#1358](https://github.com/androidx/media/issues/1358)).
 *   Transformer:
     *   Work around a decoder bug where the number of audio channels was capped
         at stereo when handling PCM input.